+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * `etcpal_thread_os_handle_t`
   * etcpal_thread_get_os_handle()
   * etcpal_thread_get_current_os_handle()
+- New C++ feature: Strongly typed opaque IDs (`etcpal/cpp/opaque_id.h`)
 
 ### Changed
 - Separated etcpal/lock.h into more specific headers: etcpal/mutex.h, etcpal/signal.h, etcpal/sem.h